more shared types
