+++ types/schemas/application/data/Applicant.ts
@@ -1,5 +1,6 @@
 import {Address, UserAddress} from '../../../shared/Addresses';
 import {ContactDetails} from '../../../shared/Contacts';
+import {MaintenanceContacts} from '../../../shared/MaintenanceContact';
 import {Ownership} from '../../../shared/Ownership';
 import {SiteContact} from '../../../shared/SiteContact';
 
@@ -18,7 +19,7 @@ export type BaseApplicant = ContactDetails & {
   address: UserAddress;
   ownership?: Ownership;
   siteContact: SiteContact;
-  maintenanceContact?: MaintenanceContact;
+  maintenanceContact?: MaintenanceContacts;
 };
 
 /**
@@ -28,16 +29,3 @@ export type BaseApplicant = ContactDetails & {
 export interface Agent extends BaseApplicant {
   agent: ContactDetails & {address: Address};
 }
-
-/**
- * @id #MaintenanceContact
- * @description Contact information for the person(s) responsible for maintenance while the works are carried out
- */
-export type MaintenanceContact = {
-  when:
-    | 'duringConstruction'
-    | 'afterConstruction'
-    | 'duringAndAfterConstruction';
-  address: Address;
-  contact: ContactDetails;
-}[];
