+++ types/schemas/prototypeApplication/data/Property.ts
@@ -9,6 +9,7 @@ import {
   PlanningDesignation,
 } from '../enums/PlanningDesignation';
 import {PropertyType} from '../enums/PropertyTypes';
+import {ExistingLondonParking} from '../../../shared/Parking';
 
 export type PropertyBase = EnglandProperty | LondonProperty;
 
@@ -65,17 +66,7 @@ export interface LondonProperty extends EnglandProperty {
       | 'No';
     number?: string;
   };
-  parking?: {
-    cars?: ExistingCount;
-    vans?: ExistingCount;
-    motorcycles?: ExistingCount;
-    cycles?: ExistingCount;
-    buses?: ExistingCount;
-    disabled?: ExistingCount;
-    carClub?: ExistingCount;
-    offStreet?: {residential: ExistingCount};
-    other?: ExistingCount;
-  };
+  parking?: ExistingLondonParking;
 }
 
 type ExistingCount = {count: number};
