+++ types/schemas/prototypeApplication/data/Proposal.ts
@@ -1,5 +1,6 @@
 import {GeoBoundary} from '../../../shared/Boundaries';
 import {Materials} from '../../../shared/Materials';
+import {ProposedLondonParking} from '../../../shared/Parking';
 import {Area, Date, Integer} from '../../../shared/utils';
 import {PrimaryApplicationType} from '../enums/ApplicationType';
 import {BuildingRegulation} from '../enums/BuildingRegulation';
@@ -138,20 +139,7 @@ export interface EnglandProposal {
  */
 export interface LondonProposal extends Omit<EnglandProposal, 'units'> {
   schemeName?: string;
-  /**
-   * @description Proposed parking spaces
-   */
-  parking?: {
-    cars?: ProposedCount;
-    vans?: ProposedCount;
-    motorcycles?: ProposedCount;
-    cycles?: ProposedCount;
-    buses?: ProposedCount;
-    disabled?: ProposedCount;
-    carClub?: ProposedCount;
-    offStreet?: {residential: ProposedCount};
-    other?: ProposedCount;
-  };
+  parking?: ProposedLondonParking;
   /**
    * @description Creating new buildings
    */
@@ -262,11 +250,6 @@ export interface LondonProposal extends Omit<EnglandProposal, 'units'> {
     };
   };
 }
-
-type ProposedCount = {
-  count: number;
-  difference: number;
-};
 
 /**
  * @description Details about creating new buildings or increasing the height of existing buildings
